docs: improve metadata description verbiage
The text as written is ambiguous about if the directives _only_ tag things with Crypto or with the icon, and it's not until the reader scrolls down that they might figure out the reference is to the example below.
Adding this small bit of text directs them to the example early on and adds clarity.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>
Co-authored-by: Adrian Goins <171292+oskapt@users.noreply.github.com>